Refactor record data validation and sorting
Refactored the records_by_round_view function to validate and sort records data using a new function called validate_and_sort_records. This function uses the Pydantic model for validation and then sorts the records. To better represent its function, the name of the method custom_representation in the Record class was changed to template_representation.